fix 路线图适配，文物预览 add AR等

+++ SilkRoad/SilkRoad/Classes/Business/VR/Model/RoadMapCityLocation.swift
@@ -9,8 +9,8 @@ import Foundation
 import UIKit
 
 struct RoadMapCityLocation {
-    var minX: Int
-    var minY: Int
-    var maxX: Int
-    var maxY: Int
+    var minX: Float
+    var minY: Float
+    var maxX: Float
+    var maxY: Float
 }
